added

+++ WebApplication1/WebApplication1/Controllers/InvoiceController.cs
@@ -33,7 +33,7 @@ namespace WebApplication1.Controllers
                         while (reader.Read())
                         {
                             Invoice invoiceList = new Invoice();
-                            invoiceList.Invoice_Head_id = reader.GetInt32(0);
+                            invoiceList.Invoice_Head_id = reader.GetInt16(0);
                             invoiceList.Invoice_Head_Customer = reader.GetString(1);
                             invoiceList.Invoice_Head_Amount = reader.GetString(2);
                             invoiceList.Invoice_Head_Date = reader.GetDateTime(3);
